backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
textblob>=0.17.1
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
torch>=2.0.0
bcrypt>=4.0.1
requests>=2.31.0
beautifulsoup4>=4.12.2
nltk>=3.8.1
